fix x64 code build

+++ src/Frameworks/PythonFramework/PythonScriptModuleFinder.h
@@ -60,13 +60,13 @@ namespace Mengine
     protected:
         bool find_module_source_( pybind::kernel_interface * _kernel, PyObject * _module, const ScriptModuleLoaderPtr & _loader );
         bool find_module_code_( pybind::kernel_interface * _kernel, PyObject * _module, const ScriptModuleLoaderPtr & _loader );
-        bool find_module_( pybind::kernel_interface * _kernel, PyObject * _module, const ScriptModuleLoaderPtr & _loader, const Char * _ext, uint32_t _extN, const Char * _init, uint32_t _extI );
+        bool find_module_( pybind::kernel_interface * _kernel, PyObject * _module, const ScriptModuleLoaderPtr & _loader, const Char * _ext, size_t _extN, const Char * _init, size_t _extI );
 
     protected:
-        bool convertDotToSlash_( pybind::kernel_interface * _kernel, Char * const _cache, uint32_t _cacheSize, PyObject * _module, uint32_t * const _modulePathCacheLen );
+        bool convertDotToSlash_( pybind::kernel_interface * _kernel, Char * const _cache, size_t _cacheSize, PyObject * _module, size_t * const _modulePathCacheLen );
 
     protected:
-        bool findModule_( const Char * _modulePath, uint32_t _modulePathLen, const ScriptModuleLoaderPtr & _loader ) const;
+        bool findModule_( const Char * _modulePath, size_t _modulePathLen, const ScriptModuleLoaderPtr & _loader ) const;
 
     protected:
         pybind::object m_embed;
